requirements.txt
aiohttp==3.8.2
yarl==1.8.1
frozenlist==1.3.1
discord.py